+++ CHANGELOG.md
@@ -3,8 +3,7 @@
 ### 1.7.6
 * `ThumbCreator` can throw a `BadMethodCallException` or a `NotWritableException`;
 * `ThumbHelper` throws a `BadMethodCallException` if a method doesn't exist;
-* fixed little bug for `ThumbsPathTrait::getPath()` method;
-* little fixes.
+* fixed little bug for `ThumbsPathTrait::getPath()` method.
 
 ### 1.7.5
 * fixed little bug for `ThumbsPathTrait::resolveFilePath()` method;